+++ www/go/core/cron/GarbageCollection.php
@@ -41,8 +41,22 @@ class GarbageCollection extends CronJob {
 
 		Token::collectGarbage();
 		OauthAccessToken::collectGarbage();
+		$this->tmpFiles();
 
 		$this->fireEvent(self::EVENT_RUN);
+	}
+
+	private function tmpFiles() {
+		$garbage =go()->getTmpFolder()->find(
+			[
+				'older' => new DateTime("-1 day"),
+				'empty' => true
+			]
+		);
+
+		foreach($garbage as $item) {
+			$item->delete();
+		}
 	}
 
 	private function blobs() {