+++ www/go/modules/community/dokuwiki/language/nb.php
@@ -1,7 +1,7 @@
 <?php
 return array (
   'name' => 'DokuWiki',
-  'description' => 'Denne modulen integrerer DokuWiki med Group-Office, og logger automatisk inn Group-Office brukere. Se modules/dokuwiki/INSTALL.TXT for mer informasjon.',
+  'description' => 'Denne modulen integrerer DokuWiki med Group-Office, og logger automatisk inn Group-Office brukere. Se go/modules/community/dokuwiki/INSTALL.TXT for mer informasjon.',
   'externalUrl' => 'DokuWiki URL',
   'title' => 'Tittel',
 );