+++ www/go/modules/community/dokuwiki/language/nl.php
@@ -1,7 +1,7 @@
 <?php
 return array (
   'name' => 'Dokuwiki',
-  'description' => 'Deze module integreerd Dokuwiki in Group-Office en zorgt ervoor dat u automatisch ingelogd wordt. Bekijk modules/dokuwiki/INSTALL.TXT voor meer informatie.',
+  'description' => 'Deze module integreerd Dokuwiki in Group-Office en zorgt ervoor dat u automatisch ingelogd wordt. Bekijk go/modules/community/dokuwiki/INSTALL.TXT voor meer informatie.',
   'externalUrl' => 'Dokuwiki url',
   'title' => 'Dokuwiki',
 );